return a new Factors object to prevent bug
Earlier, the Factors object that gets returned from get_factors was
a reference to an old one. Looping through it while removing elements
from it results in a bug. A solution is to create a shallow copy by
sending a new Factors object so that we break dependence -- looping
through the new Factors object makes looping become independent of
removing factors from the old Factors object.