improve correctness of the logic.
Using topological sort for running the transition, we guaranteed that
the parents of a variable (deterministic or random) get transitioned first.
Once the parents get transitioned, the deterministic variables can then
evaluate. If those are parents of a random variable, then the random
variable uses the transitioned values as parents.